Add SepParserVector512NrwCmpExtMsbTzcnt and refactor mask to use nuint etc.

+++ src/Sep.Test/SepParseMaskTest.cs
@@ -11,7 +11,7 @@ public partial class SepParseMaskTest
     const char Separator = ';';
     const int CharsIndexOffset = 17;
 
-    static int MaskFor(ReadOnlySpan<char> chars)
+    static uint MaskFor(ReadOnlySpan<char> chars)
     {
         var mask = 0;
         for (var i = 0; i < Math.Min(chars.Length, 32); i++)
@@ -21,7 +21,7 @@ public partial class SepParseMaskTest
                 mask |= 1 << i;
             }
         }
-        return mask;
+        return (uint)mask;
     }
 
     static bool IsSpecialChar(char c) => c switch
@@ -36,7 +36,7 @@ public partial class SepParseMaskTest
     static void AssertParseState(int[] expected,
         Span<int> colEnds, ref int start, ref int end,
         int expectedRowLineEndingOffset, int rowLineEndingOffset,
-        int expectedQuoting, int quoting,
+        nuint expectedQuoting, nuint quoting,
         int expectedLineNumber, int lineNumber)
     {
         var count = (int)Unsafe.ByteOffset(ref start, ref end) / sizeof(int);
